+++ CHANGELOG.md
@@ -52,6 +52,7 @@ speaking, considered to be the player's experience.
   of units.
 - Fixed the corrupted display of a Mayan Town name, a Norman unit name, and a
   Greek unit name in the Esperanto, Polish, and Russian translations.
+- Fixed the mouse cursor appearing under instead of over some text.
 - Fixed the alignment of the town migration button (pink cross with arrows in a
   box) not being in the centre.
 - Fixed the calculation of the nearest Market, which sometimes resulted in